requirements.txt
aloe==0.2.0
aloe-django==0.2.0
ansicolors==1.1.8
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cssselect==1.1.0
Django==3.1.4
django-filter==2.4.0
django-nose==1.4.7
djangorestframework==3.12.2
docopt==0.6.2
gherkin-official==4.1.3
idna==2.10
Markdown==3.3.3
nose==1.3.7
numpy
pandas
prettytable==2.1.0
psutil==5.8.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tabulate==0.8.9
urllib3==1.26.4
uWSGI==2.0.19.1
wcwidth==0.2.5
yarg==0.1.9

